+++ package.json
@@ -28,11 +28,11 @@
     "interweave": "^12.7.2",
     "lodash": "^4.17.21",
     "metamask-react": "^1.3.1",
-    "next": "^10.2.3",
+    "next": "^11.0.0",
     "opensea-js": "^1.1.11",
-    "react": "17.0.2",
+    "react": "^17.0.2",
     "react-bootstrap": "^1.5.2",
-    "react-dom": "17.0.2",
+    "react-dom": "^17.0.2",
     "react-elastic-carousel": "^0.11.5",
     "react-multi-carousel": "^2.6.2",
     "react-query": "^3.16.0",
