--- webpack.config.js
@@ -1,5 +1,0 @@
-// plugins: [
-//   new webpack.ProvidePlugin({
-//     React: "react",
-//   }),
-// ];
